Lazily instantiate database connection per request
If we open the connection at application startup like you might normally
do with a database connection, then Serverless Aurora doesn't scale down
because there's always a persistent connection.

The overhead of instantiating a new connection per request is going to
be fine for this project, relative to how much it would cost otherwise.

We don't need to modify `blast` because it already only connects once a
day to run.